DefinitionItem and ParagraphLink to separate files
